Update README.md
@dacb-bu
This commit includes some potentially helpful information for future people working on the site to use to better understand the search engine. It also explains to potential future developers how the code is formatted and designed. This should help in making this search system much more accessible for potential future developers on the site, so that they could get started and make changes as fast and as effectively as possible.